refactor: create route group home
