+++ src/app/(home)/home.module.css
@@ -1,0 +1,55 @@
+.main {
+  align-items: center;
+  background-color: var(--clr-b);
+  display: flex;
+  justify-content: space-evenly;
+  min-height: 100vh;
+}
+
+.main > a {
+  aspect-ratio: 1;
+  display: flex;
+  left: 20px;
+  position: absolute;
+  top: 20px;
+  width: 30px;
+}
+
+.section {
+  --r: 16px;
+
+  border-radius: var(--r);
+  box-shadow: 0 2px 4px 0 rgb(0 0 0 / 20%);
+  display: flex;
+  flex-direction: column;
+  width: 420px;
+}
+
+.section > div {
+  align-items: center;
+  background-color: white;
+  border-radius: 0 0 var(--r) var(--r);
+  display: flex;
+  flex-direction: column;
+  height: 600px;
+  justify-content: center;
+}
+
+.section > header {
+  align-items: center;
+  background-color: var(--clr-d);
+  border-radius: var(--r) var(--r) 0 0;
+  color: white;
+  display: flex;
+  height: 120px;
+  justify-content: space-between;
+  padding: 0 40px;
+}
+
+.section > header > h2 {
+  font-size: 2rem;
+}
+
+.section > header > span {
+  font-size: 1.5rem;
+}
