--- src/components/styled/Main.styled.ts
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-export const Main = styled.main`
-  align-items: center;
-  background-color: var(--clr-b);
-  display: flex;
-  justify-content: space-evenly;
-  min-height: 100vh;
-`;
